commit
ahhh the enemy cant deal damage to the player at the moment, but the player can self damage.

+++ Assets/Script/Stats/Stat.cs
@@ -8,9 +8,9 @@ public class Stat
     //serialized class for handling Stats
 
     [SerializeField]
-    private int baseValue;
+    private float baseValue;
 
-    public int getValue()
+    public float getValue()
     {
         return baseValue;
     }
